Let list action use a match/filter method
The cli decides which content match/filter to use
(currently 'match all' and 'match only the name specified')
and uses it when listing to select the subset to show.

That means 'mazer list' now can show all content,
a single content, or multiple contents specified on the
command line.

For example:

   'mazer list alikins.some-role'
    (show only that role)

   'mazer list'
   (show all)

   'mazer list alikins.some-role alikins.awx'
   (show only alikins.some-role and alikins.awx)

Previously only one cli arg was used.