Remove support for ns.name.sub_name in content_spec
The 'sub_name' was the potential third part of a dot
separated content_spec_string that previously selected
which subset of a galaxy content to install (for ex,
a particular module or role).

That is no longer supported, so remove it from the
content_spec parsing and update the unit tests.