Create and send a X-Request-ID on http requests.
On each http request to galaxy, create a random (uuid4()) X-Request-ID
and include it as a header in requests. Also log the X-Request-ID when
logging the info about http requests made.

See https://blog.heroku.com/http_request_id_s_improve_visibility_across_the_application_stack
for an explanation of X-Request-ID.

The X-Request-ID allows tracing of a specific client http request to
galaxy. Galaxy includes the X-Request-ID in any log events emitted when
handling that request. That makes it much easier to troubleshoot
specific errors and failures across both the mazer client and the galaxy
server.

Note that the value is random, and is unique to each http request.