Add file collectors to  'find files to include in collection artifact'
Add a genericish FileWalker.

FileWalker is a sort of DWIM file tree walker for a
collection. It ignores .git/, *.pyc, __pycache__, *.retry
but doesn't pay attention to .gitignore and doesn't need
to run against a git managed directory.

Note: there is also currently a walker that uses setuptools_scm's git support.
But currently will only include files that are in a git
repo and assumes collection_root is a git repo.

ie, it will ignore locally generated files.
Also wont work on anything thats not git.