wip, drive install with list of ContentSpec not GalaxyContent
_build_content_spec_list (prev _build_content_set) not
returns a list of ContentSpec objects instead of instances
of GalaxyContent

plumb that through rest of install

Remove the check that uses the GalaxyContent instance to
check if it is installed (now in stuff_for_updating()).
Hoping the semantics of archive.extract_files() will do
the right thing for cleaning up old installs but it
probably will not.

Pretty much need some level of a ContentTransaction at
this point, so 'installs' that are actually updates
become a paired delete/create. And/or make sure the
archive extraction handling has a concept of 'update'
and track that transaction operation for each content
we are going to install.