Let 'already installed' detection wait un install_repositories
Instead of raising exception on the first thing that is already
installed up front, just warn about it in install_repositories
(but before finding/fetching the repo from remote).

install_repositories will just ignore it if is already installed
so that:

        mazer installl ns.something_installed_already ns.new_thing

Will warn that 'ns.something_installed_already' is in fact already
installed. But also continue and install ns.new_thing.