Load spdx data from json on first use only
Generate a short concise ansible_galaxy/data/spdx_license.json
from the verbose upstream data.

ansible_galaxy.data.spdx_license is a module that caches the
data loaded from json on first use of get_spdx.

- Add tool (spdx_update.py) to build a condensed json file
with data from spdx.json

Use this to parse spdx json data when it is updated
and generated a new source file.

- Add a 'dev/spdx-update' make target

This target pulls down the latest SPDX license
info from https://github.com/spdx/license-list-data
to data/spdx_licenses.json

Then use spdx_update.py to generate the condensed
ansible_galaxy/data/spdx_licenses.json based
on data/spdx_licenses.json

- Validate CollectionInfo off static dict of spdx info.

Previously everytime a CollectionInfo() was created, the
spdx_data was being loaded and parsed.

Now check it against a set of data in ansible_galaxy.data.spdx_licenses

If a license id is in the SPDX_LICENSES dict, it is acceptable.
If the license items value is True, then it is deprecated and
a warning is printed.

TODO: mv the deprecated license check so it only runs on 'build'
and not anytime a collection is loaded