Use factory=list instead of default=[] for attr default.
authors = attr.ib(default=[]) makes the default value for
the 'authors' attribute of the class a mutable list. It
can be modified and later new instances will use the mutated
list as the default instead of new empty list.

Use attr.ib's factory option to specify to use an empty list
if otherwise not provided.