Assorted changes to support semver specs/ranges
repository_version.get_repository_version now takes in
a RequirementSpec instead of a version string. This allows
that to be a spec/range, but it also means that 'v1.0.0'
style aliases need tweaks. ala RequirementSpec.version_aka

If we end up with two ambiquous versions ('1.0.0',
and 'v1.0.0') that both match the RequirementSpec,
now we just throw an exception instead of trying to
keep track of which is which.

Update the repository_version tests.