Top level exc message get logged to stderr_log as well
Currently stderr_log is plain output to stderr, but
could add formatting/color/etc in the future.

The distinction between stderr_log and the default
display_callback is that stderr_log is the last ditch
output for the cli main(), while display_callback is called
from within ansible_galaxy library and could potentially be
console or a GUI or a log, etc.