Add license_file to col_info, make 'license' a list
Add unit tests. And some notes/examples for the regexes.

'license' is now a list, and single items are converted to
a single item list if provided.

Change validation of license identifiers to check all identifiers
and confirm they are SPDX identifiers. If there is one or more invalid
license identifiers, a ValueError is raised with an error message
listing all of the invalid license identifiers.

Note: This allows license=None and license=[] in CollectionInfo()