Simplify InstallDestinationInfo, collections_path
InstallDestinationInfo no longer needs to track separate
info about where it was told to install and where it
needs to extract the archive to. Previously this was
required because trad roles would be extracted into
a subdir of namespace/name/ (roles/...).

Since the extract_archive_to_dir is now predictable
based on only the collections_path, namespace, and name
we dont need to specify it, so remove the attribute
on InstallDestinationInfo