[WIP] debug logging, debugging scm_url installs
post install/extract of a scm url 'archive', when we
search the installed repos for the the requested package
to verify it was installed, the repo spec for the scm url
doesn't match the repo spec of the installed collection.

Some of that is because the installed version is None,
and the requirement spec version_spec was '*'.