Use 'artifact' data from API and return via fetcher
Add 'artifact' key to dict returned fromFetch.fetch()

Add 'artifact' to fetch results. The value of the 'artifact'
item in the fetch results dict should be a dict.

The dict should contain 'sha256', 'filename', and 'size'.

The data is based on the ArtifactDetail from the Galaxy REST
API. ie, results of

 GET /api/v2/collections/{namespace}/{name}/versions/{version}/artifact/

Or the 'artifact' field of:
  GET /api/v2/collections/{namespace}/{name}/versions/{version}